re-separated CartDAO and UserDAO, functionally tested but unit tests not yet written

+++ estore-api/src/main/java/com/estore/api/estoreapi/model/UserAccount.java
@@ -28,7 +28,7 @@ public class UserAccount {
     }
 
     public int getId() {
-        return this.id;
+        return this.id.intValue();
     }
 
     public String toString() {
